src/requirements.txt
Django==3.2.7	
django-environ==0.7.0
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.2.1
djangorestframework==3.13.1
django-prettyjson
djangorestframework-simplejwt==4.8.0
django-split-settings
django-phonenumber-field==5.2.0
django-cors-headers==3.13.0
django-redis==5.2.0
phonenumbers==8.12.33
django-import-export==2.8.0
django-money==3.0.0
sentry-sdk==1.9.5
requests==2.28.1
qrcode==7.3.1
psycopg2-binary==2.9.1
pymorphy2==0.9.1
gunicorn==20.1.0
docker
Pillow==9.2.0
uritemplate==4.1.1
python-dateutil==2.8.2
django-emoji-picker==0.0.6
pydantic==1.10.2
django-debug-toolbar==3.6.0
mixer
pytest-django
pytest-mock
Django==3.2.7
django-environ==0.7.0
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.2.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
psycopg2-binary==2.9.1
flake8==3.9.2
black==21.6b0
isort==5.9.3
djoser==2.1.0
PyJWT==2.1.0
pytest-django==4.4.0
pytest-factoryboy==2.1.0
Faker==8.9.1
pytest-cov==2.12.1
redis==3.5.3
celery==4.4.7
flower==0.9.7
django-celery-email==3.0.0
argh==0.26.2
PyYAML==5.3.1
watchdog==1.0.2
django-anymail==8.6
drf-yasg==1.21.4
